Use a stringInfo instead of a char for replace_string in pg_regress
The exposed function replase_string() in pg_regress assumes that there exists
enough space in the string it operates on. While this is documented and
expected, the consequence can be subtle failures in the tests which might be
hard to trace back to the root cause.

Provide an alternative function that operates on a StringInfo instead.

Co-authored-by: Asim R P <pasim@vmware.com>