Use a stringInfo instead of a char for replace_string in pg_regress
The exposed function replace_string() in pg_regress assumes that there exists
enough space in the string it operates on. While this is documented and
expected, the consequence can be subtle failures in the tests which might be
hard to trace back to the root cause.

Now replace_string() operates in StringInfo instead, growing the string as
nessecary.

Also, pg_regress takes advantage of the growing capability of StringInfo and
consumes lines greater than the original buffer, guarding against cutting the
replace string along consecutive reads.

Refactor ecpg's pg_regress slightly to make use of the new replace_string or
remove it all together where applicable.

Co-authored-by: Asim R P <pasim@vmware.com>