Teach pg_receivewal to use lz4 compression
The program pg_receivewal can use gzip compression to store the received WAL.
This commit teaches it to also be able to use lz4 compression. It is required
that the binary is build using the -llz4 flag. It is enabled via the --with-lz4
flag on configuration time.

Previously, the user had to use the option --compress with a value between [0-9]
to denote that gzip compression was required. This specific behaviour is
maintained. A newly introduced option --compression-method=lz4 can be used to ask
for the logs to be compressed with lz4. In that case, no compression values can
be selected as it did not seem too useful.

Under the hood there is nothing exceptional to be noted. Tar based archives have
not yet been taught to use lz4 compression. If that is felt useful, then it is
easy to be added in the future.

Tests have been added to verify the creation and correctness of the generated
lz4 files. The later is achieved by the use of lz4 program. Autoconf has been
taught to unconditionally recognize the existance of the program and propagate
the information to the tests.