Introduce pg_receivewal gzip compression tests
There exists a non trivial amount of code that handles gzip compression. The
current patch introduces tests that cover creation of gzip compressed WAL files
and the handling of the partial segments. The integrity of the compressed
files is verified as is the ability to start streaming from the correct
location regardless of compression.